tv-rec-dump: A script to dump all mp4 to the big storage on Raspberry Pi
This scirpt can be used to dump all mp4 from ER130 to the storage on
Raspberry Pi. It will attach thumb and copy all mp4 to a /dump folder
on Raspberry Pi's tvrec storage. This function is useful for we record
a program manually but not from tv-sch.sh, but we still want to copy
the mp4 to Raspberry Pi for KODI.