tv-rec-post: Change the logic of attach and detach thumb from ER130 to Raspberry pi
Raspberry Pi be connected to the normal open side of Relay, AverMedia ER130 be
conntected to the normal close side of Relay. And we use a Low level trigger Relay
as the second level power pin controller after Relay. This is for controlling
the timing of enabling the power of thumb before we attach it back to ER130. Base
on USB spec(?)'s request, the VBUS power? must be turn-off before a USB device be
attached on different machines. Otherwise that the device can not be initialed (
recognized) by ER130 or anyuother machines.

Current hardware connection between Raspberry Pi with Relay and MOSFET:
  pin 3 : D+ D- data pin        be controlled by a low-level trigger Relay
  pin 4 : power/ground pin      be controlled by a low-level trigger Relay
  pin 17 : second level power/ground pin control, be controlled by a MOSFET