tv-rec-post: retry copy routine when mp4 copy fail
Sometimes driver or filesystem problem causes that the USB drive
can not be mount success. It causes that the *.mp4 can not be found.
Which means that we saw *.tvschC file. Let's try to run copy routine
second time to cover this problem.

If the second round is still fail, then we copy *.tvschC to *.tvschE.
The states of tvrec file are:

*.tvsch file state:
	Initial -> [F]inal -> [C]opy -> [D]one/[E]rror