ER130-utils: create utilities file for controlling AverMedia ER130
Collected helper functions in ER130-utils for controlling AVerMedia
ER130 recording box. The first two functions are ledSampling() and
ledConstantlyGreenBright(). The ledSampling() is for sampling LED
states on ER130 through photoresistor. The ledConstantlyGreenBright()
is a parser for checking those samples to confirm that the LED is in
constantly green bright state. The CONSTANTGREEN variable will be set
to 'true'.