tv-rec-post: check and reset recording box
Add new step after post routines for checking the state of recording
box and reset it. In current use case, sometimes that the ER130 can
NOT remount USB AV hard drive after rebooting for unknown reason. The
fail rate is around 3.57%.

The workaround is that using a photoresistor for sampling the lighting
value of LED on ER130 then parse those samples. If the LED is NOT in
constantly green bright state. Then we reset power(aka. vbus) of USB HD
and reboot ER130 for trying to remount HD.