tv-rec-post: enable auto reset AV HD power function
Enabled auto reset AV HD power function when USB HD NOT be auto-mount
when ER130 rebooting.

In current use case, sometimes that the ER130 can NOT remount USB AV
hard drive after rebooting for unknown reason. The fail rate is around
3.57%.

When we found that the LED is NOT constant green bright. Reset the
power of USB AV HD (aka. vbus) and reboot ER130 for triggering
auto-mount again.