tv-rec-post: Checking green flashing and constantly amber bright of LED
Checking the green flashing and constantly amber bright LED states
if the first checking result is NOT constantly green bright. Which
means that ER130 does NOT auto-pick up USB disk after booting, or
the ER130 is NOT powered-on.

The power of USB disk should be reset and reboot ER130 for auto-pickiing
up disk again. The ER130 must be turned on if the LED state is constantly
amber bright.