tv-rec-post-copy: find out the mp4 file with smallest deviation
The original logic is finding out the matched mp4 with the
deviation less than 10 mintues. But ER130 has bug that the
internal clock will run slower and slower. After the deviation
bigger than 10 mintues, then the target task can not match with
mp4 file. So the matching logic be changed to find out mp4 file
with smallest deviation.